refactor(nvim): migrate to lazy.nvim plugin manager
Replaced packer.nvim with lazy.nvim for plugin management, updating all
plugin configurations to use the new lazy-loading structure. Removed
legacy packer and direct require-based plugin setups, consolidating
plugin declarations into modular lazy.nvim specs. Updated colorscheme,
LSP, DAP, and utility plugin configs for compatibility. This improves
startup performance, maintainability, and plugin management flexibility.
