+++ v2/nvim/lua/johnmutuma/core/init.lua
@@ -1,0 +1,2 @@
+require("johnmutuma.core.keymaps")
+require("johnmutuma.core.options")
