+++ v2/nvim/lua/johnmutuma/plugins/autopairs.lua
@@ -1,20 +1,18 @@
-local autopairs_ok, autopairs = pcall(require, "nvim-autopairs")
-
-if not autopairs_ok then
-    print("autopairs could not be loaded")
-    return
-end
-
-autopairs.setup({
-    check_ts = true,
-    ts_config = {
-        lua = { "string" }, -- it will not add a pair on that treesitter node
-        javascript = { "template_string" },
-        java = false, -- don't check treesitter on java
+return {
+    "windwp/nvim-autopairs",
+    event = "InsertEnter",
+    config = true,
+    opts = {
+        check_ts = true,
+        ts_config = {
+            lua = { "string" }, -- it will not add a pair on that treesitter node
+            javascript = { "template_string" },
+            java = false,       -- don't check treesitter on java
+        },
     },
-})
+}
 
 -- If you want insert `(` after select function or method item
-local cmp_autopairs = require("nvim-autopairs.completion.cmp")
-local cmp = require("cmp")
-cmp.event:on("confirm_done", cmp_autopairs.on_confirm_done())
+-- local cmp_autopairs = require("nvim-autopairs.completion.cmp")
+-- local cmp = require("cmp")
+-- cmp.event:on("confirm_done", cmp_autopairs.on_confirm_done())
