+++ v2/nvim/lua/johnmutuma/plugins/indent-blankline.lua
@@ -8,7 +8,13 @@ globl.indentLine_char = "│"
 opt.list = true
 -- opt.listchars:append("eol:↴")
 
-require("ibl").setup({
-    -- show_end_of_line = true,
-    scope = { enabled = false },
-})
+return {
+    "lukas-reineke/indent-blankline.nvim",
+    event = { "BufReadPre", "BufNewFile" },
+    config = function()
+        require("ibl").setup({
+            -- show_end_of_line = true,
+            scope = { enabled = false },
+        })
+    end,
+}
