+++ v2/nvim/lua/johnmutuma/plugins/lsp/lsp-lens.lua
@@ -1,20 +1,19 @@
-local success, lsp_lens = pcall(require, "lsp-lens")
-
-if not success then
-    print("lsp-lens failed to load")
-    return
-end
-
-lsp_lens.setup({
-    enable = true,
-    include_declaration = false, -- Reference include declaration
-    sections = { -- Enable / Disable specific request, formatter example looks 'Format Requests'
-        definition = true,
-        references = true,
-        implements = false,
-        git_authors = false,
-    },
-    ignore_filetype = {
-        "prisma",
-    },
-})
+return {
+    "VidocqH/lsp-lens.nvim",
+    event =  "LspAttach",
+    config = function()
+        require("lsp-lens").setup({
+            enable = true,
+            include_declaration = false, -- Reference include declaration
+            sections = {                 -- Enable / Disable specific request, formatter example looks 'Format Requests'
+                definition = true,
+                references = true,
+                implements = false,
+                git_authors = false,
+            },
+            ignore_filetype = {
+                "prisma",
+            },
+        })
+    end,
+}
