--- v2/nvim/lua/johnmutuma/plugins/nerdtree.lua
@@ -1,8 +1,0 @@
-local globl = vim.g
-local keymap = vim.keymap
-
-keymap.set("n", "<C-n>", ":NERDTreeFind<CR>:<CR>$")
-
-globl.NERDTreeWinSize = 60
-
-return {}
