+++ v2/nvim/lua/johnmutuma/plugins/nvim-tree.lua
@@ -1,3 +1,4 @@
+local vim = vim
 local globl = vim.g
 local keymap = vim.keymap
 
@@ -6,54 +7,53 @@ globl.loaded_netrw = 1
 globl.loaded_netrwPlugin = 1
 keymap.set("n", "<C-n>", ":NvimTreeToggle<CR>:<CR>$")
 
--- settings or options
+return {
+    "nvim-tree/nvim-tree.lua",
+    dependencies = {
+        "nvim-tree/nvim-web-devicons", -- optional, for file icons
+    },
+    event = { "VeryLazy" },
+    config = function()
+        require("nvim-tree").setup({
+            filesystem_watchers = {
+                enable = false,
+                debounce_delay = 50,
+                ignore_dirs = {
+                    "node_modules",
+                    "lib",
+                },
+            },
+            update_focused_file = {
+                enable = true,
+                update_root = true,
+                update_cwd = true,
+                ignore_list = { "node_modules" },
+            },
+            git = {
+                enable = true,
+                timeout = 3000,
+            },
+            filters = {
+                git_ignored = true,
+            },
+            view = {
+                width = 50,
+            },
+            renderer = {
+                indent_markers = {
+                    enable = true,
+                },
+            },
+        })
 
-local success, nvimTree = pcall(require, "nvim-tree")
-if not success then
-    print("nvim-tree failed to load")
-    return
-end
+        vim.cmd([[
+          :hi NvimTreeCursorLine guibg=#444548
+          :hi NvimTreeGitDirtyIcon guifg=red
+          :hi NvimTreeModifiedIcon guifg=red
+        ]])
 
-nvimTree.setup({
-    filesystem_watchers = {
-        enable = false,
-        debounce_delay = 50,
-        ignore_dirs = {
-            "node_modules",
-            "lib",
-        },
-    },
-    update_focused_file = {
-        enable = true,
-        update_root = true,
-        update_cwd = true,
-        ignore_list = { "node_modules" },
-    },
-    git = {
-        enable = true,
-        timeout = 3000,
-    },
-    filters = {
-        git_ignored = true,
-    },
-    view = {
-        width = 50,
-    },
-    renderer = {
-        indent_markers = {
-            enable = true,
-        },
-    },
-})
-
-vim.cmd([[
-  :hi NvimTreeCursorLine guibg=#444548
-  :hi NvimTreeGitDirtyIcon guifg=red
-  :hi NvimTreeModifiedIcon guifg=red
-]])
-
-if vim.opt.background:get() == "light" then
-    vim.cmd([[
-    :hi NvimTreeCursorLine guibg=#c1c1c1 guifg=#252525
-  ]])
-end
+        if vim.opt.background:get() == "light" then
+            vim.cmd([[ :hi NvimTreeCursorLine guibg=#c1c1c1 guifg=#252525 ]])
+        end
+    end,
+}
