--- v2/nvim/lua/johnmutuma/plugins/vim-fugitive.lua
@@ -1,4 +1,0 @@
--- define :Browse to enable GBrowse to work
-vim.cmd([[
-  command! -nargs=1 Browse silent execute '!open' shellescape(<q-args>,1)
-]])
